MobX State Persist
Keep the state whenever pages change using _page.js override. Server gives a full page with the initial state of the store from the last page.

A fully separate page, but it ships with the latest state. This should be SSR compatible later on as well, doing it this way.