DocumentEditor, DraftView, and EditorView all converted to hooks
Each one tested in every way I could think of. You can import DraftView directly into [doc].js for instance and it will behave as an independent editor. This is the case with all the FC's. If you want to hook them into any application, then pass in a ref as a prop and then you will have full access to all of their methods to integrate with your app.